feat(server): split core mod into base modules
