+++ server/src/main/java/net/lapidist/colony/base/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Built-in mods providing base server functionality.
+ */
+package net.lapidist.colony.base;
